backend/requirements.txt
# Core Django
Django==4.2.24  # Updated: Fixes 7 security vulnerabilities
djangorestframework==3.15.2  # Updated: Fixes XSS vulnerability
django-cors-headers==4.9.0  # Updated to latest compatible version
django-filter==25.1  # Updated to latest version

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.5.1  # Updated: Fixes authentication bypass vulnerability
django-allauth==0.57.0

# Environment variables
python-decouple==3.8

# Redis cache
redis==5.0.1
django-redis==5.4.0

# API documentation (OpenAPI 3.0)
drf-spectacular==0.27.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Security
bleach==6.1.0

# Development
django-debug-toolbar==4.2.0